requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.4.0
backports.asyncio.runner==1.2.0
bidict==0.23.1
black==25.9.0
blinker==1.9.0
blis==1.3.3
Brotli==1.1.0
catalogue==2.0.10
cattrs==25.3.0
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
ConfigArgParse==1.7.1
coverage==7.11.0
coveralls==4.0.1
croniter==6.0.0
cymem==2.0.13
datasets==4.4.1
dill==0.4.0
distlib==0.4.0
dnspython==2.8.0
docopt==0.6.2
email-validator==2.3.0
exceptiongroup==1.3.0
Faker==37.12.0
fastapi==0.118.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
frozenlist==1.8.0
fsspec==2025.10.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.15
idna==3.10
iniconfig==2.1.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.41.6
locust-cloud==1.27.6
lsprotocol==2023.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.6
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
preshed==3.0.12
propcache==0.4.1
psutil==7.1.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyarrow==22.0.0
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.1
pygls==1.3.1
Pygments==2.19.2
pytesseract==0.3.13
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-benchmark==5.1.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.2
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==6.4.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
rq==2.6.0
ruff==0.14.1
ruff-lsp==0.0.62
sentry-sdk==2.40.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlmodel==0.0.27
srsly==2.5.2
starlette==0.48.0
thinc==8.3.10
tomli==2.2.1
tqdm==4.67.1
typer==0.19.2
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uuid==1.30
uvicorn==0.37.0
uvloop==0.21.0
virtualenv==20.35.3
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.3
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==2.0.1
wsproto==1.2.0
xxhash==3.6.0
yarl==1.22.0
zope.event==6.0
zope.interface==8.0.1
opencv-python==4.11.0.86
gliner==0.2.24
supabase==2.27.1
